+++ src/slurmctld/slurmctld.h
@@ -401,7 +401,7 @@ typedef struct slurmctld_resv {
 				 * reservation to last                  */
 	time_t end_time;	/* end time of reservation		*/
 	char *features;		/* required node features		*/
-	uint32_t flags;		/* see RESERVE_FLAG_* in slurm.h	*/
+	uint64_t flags;		/* see RESERVE_FLAG_* in slurm.h	*/
 	bool full_nodes;	/* when reservation uses full nodes or not */
 	uint32_t job_pend_cnt;	/* number of pending jobs		*/
 	uint32_t job_run_cnt;	/* number of running jobs		*/
